add http trigger and move shared code around
