+++ serviceBusTestingWithContainer/processHttpInvoiceMsg/__init__.py
@@ -1,0 +1,28 @@
+import logging
+import json
+import azure.functions as func
+from ..shared.mainProc import processMain 
+
+async def main(req: func.HttpRequest) -> func.HttpResponse:
+    logging.info('Python HTTP trigger function processed a request.')
+
+    name = req.params.get('name')
+    if not name:
+        try:
+            req_body = req.get_json()
+        except ValueError:
+            pass
+        else:
+            name = req_body.get('name')
+
+    if name:
+        processMain(req_body, logging)
+        return func.HttpResponse(
+             "Success",
+             status_code=200
+        )
+    else:
+        return func.HttpResponse(
+             "Please pass a name on the query string or in the request body",
+             status_code=400
+        )
